Changed configuration file for function passing
