+++ servercfg/cfgdetails/testdata/minver_validation.txt
@@ -4,14 +4,14 @@
 LogLevelStr *string 0.7.4 log_level,omitempty
 MaxLenInLogs *int 0.7.4 max_query_len_in_logs,omitempty
 EncodeLoggedQuery *bool 0.7.4 encode_logged_query,omitempty
-BehaviorConfig *servercfg.DoltgresBehaviorConfig 0.7.4 behavior,omitempty
+BehaviorConfig *cfgdetails.DoltgresBehaviorConfig 0.7.4 behavior,omitempty
 -ReadOnly *bool 0.7.4 read_only,omitempty
 -DisableClientMultiStatements *bool 0.7.4 disable_client_multi_statements,omitempty
 -DoltTransactionCommit *bool 0.7.4 dolt_transaction_commit,omitempty
-UserConfig *servercfg.DoltgresUserConfig 0.7.4 user,omitempty
+UserConfig *cfgdetails.DoltgresUserConfig 0.7.4 user,omitempty
 -Name *string 0.7.4 name,omitempty
 -Password *string 0.7.4 password,omitempty
-ListenerConfig *servercfg.DoltgresListenerConfig 0.7.4 listener,omitempty
+ListenerConfig *cfgdetails.DoltgresListenerConfig 0.7.4 listener,omitempty
 -HostStr *string 0.7.4 host,omitempty
 -PortNumber *int 0.7.4 port,omitempty
 -ReadTimeoutMillis *uint64 0.7.4 read_timeout_millis,omitempty
@@ -21,24 +21,24 @@ ListenerConfig *servercfg.DoltgresListenerConfig 0.7.4 listener,omitempty
 -RequireSecureTransport *bool 0.7.4 require_secure_transport,omitempty
 -AllowCleartextPasswords *bool 0.7.4 allow_cleartext_passwords,omitempty
 -Socket *string 0.7.4 socket,omitempty
-PerformanceConfig *servercfg.DoltgresPerformanceConfig 0.7.4 performance,omitempty
+PerformanceConfig *cfgdetails.DoltgresPerformanceConfig 0.7.4 performance,omitempty
 -QueryParallelism *int 0.7.4 query_parallelism,omitempty
 DataDirStr *string 0.7.4 data_dir,omitempty
 CfgDirStr *string 0.7.4 cfg_dir,omitempty
-MetricsConfig *servercfg.DoltgesMetricsConfig 0.7.4 metrics,omitempty
+MetricsConfig *cfgdetails.DoltgesMetricsConfig 0.7.4 metrics,omitempty
 -Labels map[string]string 0.7.4 labels,omitempty
 -Host *string 0.7.4 host,omitempty
 -Port *int 0.7.4 port,omitempty
 -TlsCert *string 0.53.5 tls_cert,omitempty
 -TlsKey *string 0.53.5 tls_key,omitempty
 -TlsCa *string 0.53.5 tls_ca,omitempty
-RemotesapiConfig *servercfg.DoltgresRemotesapiConfig 0.7.4 remotesapi,omitempty
+RemotesapiConfig *cfgdetails.DoltgresRemotesapiConfig 0.7.4 remotesapi,omitempty
 -Port *int 0.7.4 port,omitempty
 -ReadOnly *bool 0.7.4 read_only,omitempty
 PrivilegeFile *string 0.7.4 privilege_file,omitempty
 AuthFile *string 0.54.4 auth_file,omitempty
 BranchControlFile *string 0.7.4 branch_control_file,omitempty
-Vars []servercfg.DoltgresUserSessionVars 0.7.4 user_session_vars,omitempty
+Vars []cfgdetails.DoltgresUserSessionVars 0.7.4 user_session_vars,omitempty
 -Name string 0.0.0 name
 -Vars map[string]interface{} 0.0.0 vars,omitempty
 SystemVariables map[string]interface{} 0.7.4 system_variables,omitempty
@@ -48,7 +48,7 @@ Jwks []servercfg.JwksConfig 0.7.4 jwks,omitempty
 -Claims map[string]string 0.0.0 claims
 -FieldsToLog []string 0.0.0 fields_to_log
 GoldenMysqlConn *string 0.7.4 golden_mysql_conn,omitempty
-PostgresReplicationConfig *servercfg.PostgresReplicationConfig 0.7.4 postgres_replication,omitempty
+PostgresReplicationConfig *cfgdetails.PostgresReplicationConfig 0.7.4 postgres_replication,omitempty
 -PostgresServerAddress *string 0.7.4 postgres_server_address,omitempty
 -PostgresUser *string 0.7.4 postgres_user,omitempty
 -PostgresPassword *string 0.7.4 postgres_password,omitempty
